tools/photo_cli/requirements.txt
Pillow>=10.0.0
scikit-learn>=1.4.0
numpy>=1.26.0